- (alarm.c): Patch from Michael Slifcak:   - Cast to pointer to int may overflow the return buffer.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@2328 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/v2party/alarm.c
@@ -195,7 +195,7 @@ rmonGetValue(oid *srcParty,
 	return 7;
     }
     
-    *value = *(int *)var;
+    *value = *(long *)var;
     return 0;
 }
